refactor: move icon component into dir
