+++ src/lib/components/icon/icon.module.css
@@ -1,0 +1,6 @@
+.icon {
+	aspect-ratio: 1 / 1;
+	width: var(--icon-size, 1em);
+	height: var(--icon-size, 1em);
+	fill: var(--icon-color, currentcolor);
+}
