+++ src/lib/components/icon/icon.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+	import styles from "./icon.module.css";
+
 	export let title: string | undefined = undefined;
 	export let viewBox = "0 0 256 256";
 	export let path: string | undefined = undefined;
@@ -7,7 +9,8 @@
 <svelte:options namespace="svg" />
 
 <svg
-	aria-labelledby={title ?? null}
+	aria-labelledby={title}
+	class="{styles.icon}"
 	role="presentation"
 	{viewBox}
 	{...$$restProps}
@@ -17,12 +20,3 @@
 		{/if}
 		<path d={path} />
 </svg>
-
-<style>
-	svg {
-		aspect-ratio: 1 / 1;
-		width: var(--icon-size, 1em);
-		height: var(--icon-size, 1em);
-		fill: var(--icon-color, currentcolor);
-	}
-</style>
