+++ src/lib/components/input-date/input-datetime.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import Button from "../button/button.svelte";
-	import Icon from "../icon.svelte";
+	import Icon from "../icon/icon.svelte";
 	import { iconTimeline, iconChevronLeft, iconChevronRight } from "@lib/icons";
 	import { getDateTimeString, incrementDateByDay } from "@lib/helpers";
 	import styles from "./input-datetime.module.css";
